+++ apps/website/docs/.vitepress/config.js
@@ -311,6 +311,28 @@ export default withMermaid(
               { text: 'Automated cache', link: '/recipes/cache' },
             ],
           },
+          {
+            text: 'Showcases',
+            collapsed: true,
+            items: [
+              {
+                text: 'SolidJS',
+                link: 'https://github.com/igorkamyshev/farfetched/tree/master/apps/showcase/solid-real-world-rick-morty',
+              },
+              {
+                text: 'React and React Router',
+                link: 'https://github.com/igorkamyshev/farfetched/tree/master/apps/showcase/react-real-world-pokemons',
+              },
+              {
+                text: 'Next.js',
+                link: 'https://github.com/effector/next/tree/main/apps/playground-app',
+              },
+              {
+                text: 'Forest',
+                link: 'https://github.com/igorkamyshev/farfetched/tree/master/apps/showcase/forest-real-world-breaking-bad',
+              },
+            ],
+          },
         ],
         '/statements': [
           {
@@ -336,6 +358,7 @@ export default withMermaid(
           {
             text: 'Releases',
             items: [
+              { text: 'v0.9', link: '/releases/0-9' },
               { text: 'v0.8 Saphan Hin', link: '/releases/0-8' },
               { text: 'v0.7 Nam Phu Chet Si', link: '/releases/0-7' },
               { text: 'v0.6 Huai Nam Dang', link: '/releases/0-6' },