+++ packages/core/src/mutation/create_headless_mutation.ts
@@ -1,10 +1,16 @@
+import { attach, type Store } from 'effector';
+
 import { createRemoteOperation } from '../remote_operation/create_remote_operation';
-import { DynamicallySourcedField, StaticOrReactive } from '../libs/patronus';
-import { Mutation, MutationSymbol } from './type';
-import { Contract } from '../contract/type';
-import { InvalidDataError } from '../errors/type';
-import { Validator } from '../validation/type';
-import { attach, Store } from 'effector';
+import {
+  readonly,
+  type DynamicallySourcedField,
+  type StaticOrReactive,
+} from '../libs/patronus';
+import { type Mutation, MutationSymbol } from './type';
+import { type Contract } from '../contract/type';
+import { type InvalidDataError } from '../errors/type';
+import { type Validator } from '../validation/type';
+import { type ExecutionMeta } from '../remote_operation/type';
 
 export interface SharedMutationFactoryConfig {
   name?: string;
@@ -82,9 +88,9 @@ export function createHeadlessMutation<
       attach({
         source,
         mapParams: (
-          { params, ...rest }: { params: NewParams },
+          { params, ...rest }: { params: NewParams; meta: ExecutionMeta },
           sourceValue
-        ): { params: Params } => ({
+        ): { params: Params; meta: ExecutionMeta } => ({
           params: (mapParams
             ? mapParams(params, sourceValue)
             : params) as Params,
@@ -100,7 +106,20 @@ export function createHeadlessMutation<
   // -- Public API --
 
   return {
-    ...operation,
+    start: operation.start,
+    $status: readonly(operation.$status),
+    $idle: readonly(operation.$idle),
+    $pending: readonly(operation.$pending),
+    $succeeded: readonly(operation.$succeeded),
+    $failed: readonly(operation.$failed),
+    $finished: readonly(operation.$finished),
+    $enabled: readonly(operation.$enabled),
+    finished: {
+      success: readonly(operation.finished.success),
+      failure: readonly(operation.finished.failure),
+      finally: readonly(operation.finished.finally),
+      skip: readonly(operation.finished.skip),
+    },
     __: { ...operation.__, experimentalAPI: { attach: attachProtocol } },
     '@@unitShape': unitShapeProtocol,
   };