+++ packages/core/src/retry/__tests__/retry.query.test.ts
@@ -1,3 +1,4 @@
+import { watchRemoteOperation } from '@farfetched/test-utils';
 import { allSettled, createEvent, createStore, fork } from 'effector';
 import { describe, test, vi, expect } from 'vitest';
 
@@ -130,6 +131,8 @@ describe('retry with query', () => {
             "meta": {
               "attempt": 1,
               "maxAttempts": 3,
+              "stale": false,
+              "stopErrorPropagation": false,
             },
             "params": "Initial",
           },
@@ -140,6 +143,8 @@ describe('retry with query', () => {
             "meta": {
               "attempt": 2,
               "maxAttempts": 3,
+              "stale": false,
+              "stopErrorPropagation": false,
             },
             "params": "Initial 1",
           },
@@ -150,6 +155,8 @@ describe('retry with query', () => {
             "meta": {
               "attempt": 3,
               "maxAttempts": 3,
+              "stale": false,
+              "stopErrorPropagation": false,
             },
             "params": "Initial 1 2",
           },
@@ -294,5 +301,87 @@ describe('retry with query', () => {
     await allSettled(query.start, { scope, params: 42 });
 
     expect(otherwiseListener).not.toBeCalled();
+  });
+
+  test('reset failure counter for manual query start', async () => {
+    const handler = vi.fn().mockRejectedValue(new Error('Sorry'));
+
+    const query = createQuery({
+      handler,
+    });
+
+    retry(query, { times: 1, delay: 0 });
+
+    const scope = fork();
+
+    await allSettled(query.start, { scope, params: 42 });
+
+    // 1 for start
+    // 1 for retry
+    expect(handler).toBeCalledTimes(2);
+
+    await allSettled(query.start, { scope, params: 42 });
+
+    // 1 for start
+    // 1 for retry
+    expect(handler).toBeCalledTimes(4);
+  });
+
+  test('throw error in case of retry', async () => {
+    const query = createQuery({
+      handler: vi.fn().mockRejectedValue(new Error('Sorry')),
+    });
+
+    retry(query, { times: 1, delay: 0 });
+
+    const scope = fork();
+
+    const { listeners } = watchRemoteOperation(query, scope);
+
+    await allSettled(query.start, { scope, params: 42 });
+
+    // 1 for original start
+    // 1 for retry
+    expect(listeners.onFailure).toBeCalledTimes(2);
+  });
+
+  test('throw error in case of retry with supressIntermediateErrors', async () => {
+    const query = createQuery({
+      handler: vi.fn().mockImplementation(({ attempt }) => {
+        throw new Error(`Sorry, attempt ${attempt}`);
+      }),
+    });
+
+    retry(query, {
+      times: 1,
+      mapParams({ meta }) {
+        return { attempt: meta.attempt };
+      },
+      delay: 0,
+      supressIntermediateErrors: true,
+    });
+
+    const scope = fork();
+
+    const { listeners } = watchRemoteOperation(query, scope);
+
+    await allSettled(query.start, { scope, params: { attempt: 0 } });
+
+    // 1 for retry
+    expect(listeners.onFailure).toBeCalledTimes(1);
+    expect(listeners.onFailure.mock.calls[0]).toMatchInlineSnapshot(`
+      [
+        {
+          "error": [Error: Sorry, attempt 1],
+          "meta": {
+            "stale": false,
+            "stopErrorPropagation": false,
+          },
+          "params": {
+            "attempt": 1,
+          },
+        },
+      ]
+    `);
   });
 });