+++ packages/core/src/retry/retry.ts
@@ -1,10 +1,16 @@
 import {
   combine,
+  createEffect,
   createEvent,
   createStore,
-  Event,
   sample,
   split,
+  attach,
+  scopeBind,
+  type Event,
+  type EffectError,
+  type EffectParams,
+  type EffectResult,
 } from 'effector';
 
 import {
@@ -15,26 +21,28 @@ import {
   type SourcedField,
   type StaticOrReactive,
 } from '../libs/patronus';
-import { Time, parseTime } from '../libs/date-nfs';
+import { type Time, parseTime } from '../libs/date-nfs';
 
 import {
-  RemoteOperation,
-  RemoteOperationError,
-  RemoteOperationParams,
+  type ExecutionMeta,
+  type RemoteOperation,
+  type RemoteOperationError,
+  type RemoteOperationParams,
 } from '../remote_operation/type';
-import { RetryMeta } from './type';
+import { type RetryMeta } from './type';
 
 type FailInfo<Q extends RemoteOperation<any, any, any, any>> = {
   params: RemoteOperationParams<Q>;
   error: RemoteOperationError<Q>;
+  meta: ExecutionMeta;
 };
 
-interface RetryConfig<
+type RetryConfig<
   Q extends RemoteOperation<any, any, any, any>,
   DelaySource = unknown,
   FilterSource = unknown,
   MapParamsSource = unknown
-> {
+> = {
   times: StaticOrReactive<number>;
   delay: SourcedField<RetryMeta, Time, DelaySource>;
   filter?: SourcedField<FailInfo<Q>, boolean, FilterSource>;
@@ -44,7 +52,8 @@ interface RetryConfig<
     MapParamsSource
   >;
   otherwise?: Event<FailInfo<Q>>;
-}
+  supressIntermediateErrors?: boolean;
+};
 
 export function retry<
   Q extends RemoteOperation<any, any, any, any>,
@@ -58,9 +67,11 @@ export function retry<
     delay: timeout,
     filter,
     mapParams,
-    otherwise,
+    ...params
   }: RetryConfig<Q, DelaySource, FilterSource, MapParamsSource>
 ): void {
+  const supressIntermediateErrors = params.supressIntermediateErrors ?? false;
+
   const $maxAttempts = normalizeStaticOrReactive(times);
   const $attempt = createStore(1, {
     serialize: 'ignore',
@@ -70,23 +81,36 @@ export function retry<
     attempt: $attempt,
   });
 
+  const $supressError = combine(
+    $attempt,
+    $maxAttempts,
+    (attempt, maxAttempts) =>
+      supressIntermediateErrors && attempt <= maxAttempts
+  );
+
+  const failed = createEvent<{
+    params: RemoteOperationParams<Q>;
+    error: RemoteOperationError<Q>;
+    meta: ExecutionMeta;
+  }>();
+
   const newAttempt = createEvent();
 
   const { planNextAttempt, __: retriesAreOver } = split(
     sample({
-      clock: operation.finished.failure,
+      clock: failed,
       source: {
         maxAttempts: $maxAttempts,
         attempt: $attempt,
       },
       filter: normalizeSourced({
         field: (filter ?? true) as any,
-        clock: operation.finished.failure,
+        clock: failed,
       }),
-      fn: ({ attempt, maxAttempts }, { params, error }) => ({
+      fn: ({ attempt, maxAttempts }, { params, error, meta }) => ({
         params,
         error,
-        meta: { attempt, maxAttempts },
+        meta: { ...meta, attempt, maxAttempts },
       }),
     }),
     { planNextAttempt: ({ meta }) => meta.attempt <= meta.maxAttempts }
@@ -106,14 +130,58 @@ export function retry<
         source: $meta,
       }).map(parseTime),
     }),
-    target: [newAttempt, operation.start],
+    fn: (params) => ({
+      params,
+      meta: { stopErrorPropagation: false, stale: true },
+    }),
+    target: [newAttempt, operation.__.lowLevelAPI.startWithMeta],
   });
 
   $attempt
     .on(newAttempt, (attempt) => attempt + 1)
-    .reset(operation.finished.success);
+    .reset([operation.finished.success, operation.start]);
 
-  if (otherwise) {
-    sample({ clock: retriesAreOver, target: otherwise });
+  if (params.otherwise) {
+    sample({ clock: retriesAreOver, target: params.otherwise });
   }
+
+  if (supressIntermediateErrors) {
+    const originalFx =
+      operation.__.lowLevelAPI.dataSourceRetrieverFx.use.getCurrent();
+
+    operation.__.lowLevelAPI.dataSourceRetrieverFx.use(
+      attach({
+        source: $supressError,
+        mapParams: (opts, supressError) => ({ ...opts, supressError }),
+        effect: createEffect<
+          EffectParams<
+            typeof operation.__.lowLevelAPI.dataSourceRetrieverFx
+          > & { supressError: boolean },
+          EffectResult<typeof operation.__.lowLevelAPI.dataSourceRetrieverFx>,
+          EffectError<typeof operation.__.lowLevelAPI.dataSourceRetrieverFx>
+        >(async ({ supressError, ...opts }) => {
+          const boundFailed = scopeBind(failed, { safe: true });
+          try {
+            const result = await originalFx(opts);
+
+            return result;
+          } catch (error: any) {
+            if (supressError) {
+              boundFailed({
+                params: opts.params,
+                error: error.error,
+                meta: opts.meta,
+              });
+
+              throw { error: error.error, stopErrorPropagation: true };
+            } else {
+              throw error;
+            }
+          }
+        }),
+      })
+    );
+  }
+
+  sample({ clock: operation.finished.failure, target: failed });
 }