+++ packages/core/src/trigger_api/__test__/keep_fresh.auto.test.ts
@@ -1,4 +1,4 @@
-import { allSettled, createStore, fork } from 'effector';
+import { allSettled, createStore, createWatch, fork } from 'effector';
 import { describe, test, expect, vi } from 'vitest';
 import { unknownContract } from '../../contract/unknown_contract';
 import { createJsonQuery } from '../../query/create_json_query';
@@ -212,6 +212,69 @@ describe('keepFresh, automatically', () => {
     allSettled($market, { scope, params: 'ge' });
 
     await allSettled(scope);
+
+    expect(listener).toBeCalledTimes(2);
+  });
+
+  test('does not use sources while disabled as source for refresh', async () => {
+    const $url = createStore('https://api.salo.com');
+
+    const query = createJsonQuery({
+      enabled: $url.map((url) => url.length > 0),
+      request: {
+        method: 'GET',
+        url: $url,
+      },
+      response: { contract: unknownContract },
+    });
+
+    keepFresh(query, { automatically: true });
+
+    const scope = fork({
+      handlers: [[query.__.executeFx, vi.fn(async () => 42)]],
+    });
+
+    const listener = vi.fn();
+
+    createWatch({ unit: query.refresh, fn: listener, scope });
+
+    await allSettled(query.refresh, { scope });
+
+    await allSettled($url, { scope, params: '' });
+    await allSettled($url, { scope, params: 'https://api.salo.com' });
+
+    expect(listener).toBeCalledTimes(1);
+  });
+
+  test('check source after enabling', async () => {
+    const $url = createStore('https://api.salo.com');
+    const $enabled = createStore(true);
+
+    const query = createJsonQuery({
+      enabled: $enabled,
+      request: {
+        method: 'GET',
+        url: $url,
+      },
+      response: { contract: unknownContract },
+    });
+
+    keepFresh(query, { automatically: true });
+
+    const scope = fork({
+      handlers: [[query.__.executeFx, vi.fn(async () => 42)]],
+    });
+
+    const listener = vi.fn();
+
+    createWatch({ unit: query.refresh, fn: listener, scope });
+
+    await allSettled(query.refresh, { scope });
+    expect(listener).toBeCalledTimes(1);
+
+    await allSettled($enabled, { scope, params: false });
+    await allSettled($url, { scope, params: 'https://api.v2.salo.com' });
+    await allSettled($enabled, { scope, params: true });
 
     expect(listener).toBeCalledTimes(2);
   });