+++ packages/core/src/trigger_api/__test__/keep_fresh.test-d.ts
@@ -1,3 +1,4 @@
+import { Event } from 'effector';
 import { describe, test } from 'vitest';
 
 import { Query } from '../../query/type';
@@ -10,5 +11,15 @@ describe('keepFresh', () => {
     keepFresh(q, { automatically: true });
     keepFresh(q, { triggers: [] });
     keepFresh(q, { automatically: true, triggers: [] });
+  });
+
+  test('supports any Event as trigger', () => {
+    const q = {} as unknown as Query<number, string, string>;
+
+    const voidEvent = {} as unknown as Event<void>;
+    const stringEvent = {} as unknown as Event<string>;
+
+    keepFresh(q, { triggers: [stringEvent, voidEvent] });
+    keepFresh(q, { automatically: true, triggers: [stringEvent, voidEvent] });
   });
 });